Refactor: python, cts: Move SimulStart into the _cts module.
No code changes.
